requirements.txt
transformers>=4.42.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
torch>=2.1.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0
python-dotenv>=1.0.1
streamlit>=1.34.0
